+++ sdk/hdinsight/Azure.ResourceManager.HDInsight/tests/Scenario/HDInsightApplicationTests.cs
@@ -29,8 +29,8 @@ namespace Azure.ResourceManager.HDInsight.Tests
             //Due to the HDInsightManagementTestBase's previous update, vnet restricts the cluster name have the unique first six characters.
             string clsusterNamePrefix = "G" + rgName.Substring(DefaultResourceGroupPrefix.Length) + "Cluster";
             string clusterName = Recording.GenerateAssetName(clsusterNamePrefix);
-            string storageAccountName = Recording.GenerateAssetName("ykstorageforcluster");
-            string containerName = Recording.GenerateAssetName("ykcontainer");
+            string storageAccountName = Recording.GenerateAssetName("azstorageforcluster");
+            string containerName = Recording.GenerateAssetName("container");
             _applicationName = Recording.GenerateAssetName("application");
             _scriptActionName = Recording.GenerateAssetName("InstallHue");
             var resourceGroup = await CreateResourceGroup(rgName);